--- cmd/admin-idp-ldap-info.go
@@ -1,54 +1,0 @@
-// Copyright (c) 2015-2022 MinIO, Inc.
-//
-// This file is part of MinIO Object Storage stack
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program.  If not, see <http://www.gnu.org/licenses/>.
-
-package cmd
-
-import (
-	"github.com/minio/cli"
-	"github.com/minio/madmin-go/v2"
-)
-
-var adminIDPLdapInfoCmd = cli.Command{
-	Name:         "info",
-	Usage:        "get LDAP IDP server configuration info",
-	Action:       mainAdminIDPLDAPInfo,
-	Before:       setGlobalsFromContext,
-	Flags:        globalFlags,
-	OnUsageError: onUsageError,
-	CustomHelpTemplate: `NAME:
-  {{.HelpName}} - {{.Usage}}
-
-USAGE:
-  {{.HelpName}} TARGET
-
-FLAGS:
-  {{range .VisibleFlags}}{{.}}
-  {{end}}
-EXAMPLES:
-  1. Get configuration info on the LDAP IDP configuration.
-     {{.Prompt}} {{.HelpName}} play/
-`,
-}
-
-func mainAdminIDPLDAPInfo(ctx *cli.Context) error {
-	if len(ctx.Args()) != 1 {
-		showCommandHelpAndExit(ctx, 1)
-	}
-
-	var cfgName string = madmin.Default
-	return adminIDPInfo(ctx, false, cfgName)
-}
