+++ cmd/auto-complete.go
@@ -361,30 +361,25 @@ var completeCmds = map[string]complete.Predictor{
 	"/admin/profile/start": aliasCompleter,
 	"/admin/profile/stop":  aliasCompleter,
 
-	"/admin/idp/set":    aliasCompleter,
-	"/admin/idp/info":   aliasCompleter,
-	"/admin/idp/list":   aliasCompleter,
-	"/admin/idp/remove": aliasCompleter,
+	"/idp/openid/add":     aliasCompleter,
+	"/idp/openid/update":  aliasCompleter,
+	"/idp/openid/remove":  aliasCompleter,
+	"/idp/openid/list":    aliasCompleter,
+	"/idp/openid/info":    aliasCompleter,
+	"/idp/openid/enable":  aliasCompleter,
+	"/idp/openid/disable": aliasCompleter,
 
-	"/admin/idp/openid/add":     aliasCompleter,
-	"/admin/idp/openid/update":  aliasCompleter,
-	"/admin/idp/openid/remove":  aliasCompleter,
-	"/admin/idp/openid/list":    aliasCompleter,
-	"/admin/idp/openid/info":    aliasCompleter,
-	"/admin/idp/openid/enable":  aliasCompleter,
-	"/admin/idp/openid/disable": aliasCompleter,
+	"/idp/ldap/add":     aliasCompleter,
+	"/idp/ldap/update":  aliasCompleter,
+	"/idp/ldap/remove":  aliasCompleter,
+	"/idp/ldap/list":    aliasCompleter,
+	"/idp/ldap/info":    aliasCompleter,
+	"/idp/ldap/enable":  aliasCompleter,
+	"/idp/ldap/disable": aliasCompleter,
 
-	"/admin/idp/ldap/add":     aliasCompleter,
-	"/admin/idp/ldap/update":  aliasCompleter,
-	"/admin/idp/ldap/remove":  aliasCompleter,
-	"/admin/idp/ldap/list":    aliasCompleter,
-	"/admin/idp/ldap/info":    aliasCompleter,
-	"/admin/idp/ldap/enable":  aliasCompleter,
-	"/admin/idp/ldap/disable": aliasCompleter,
-
-	"/admin/idp/ldap/policy/entities": aliasCompleter,
-	"/admin/idp/ldap/policy/attach":   aliasCompleter,
-	"/admin/idp/ldap/policy/detach":   aliasCompleter,
+	"/idp/ldap/policy/entities": aliasCompleter,
+	"/idp/ldap/policy/attach":   aliasCompleter,
+	"/idp/ldap/policy/detach":   aliasCompleter,
 
 	"/admin/policy/info":     aliasCompleter,
 	"/admin/policy/update":   aliasCompleter,
